+++ dotnet/Models/RatingResponse.cs
@@ -10,6 +10,21 @@ namespace ReviewsRatings.Models
         [JsonProperty("average")]
         public decimal Average { get; set; }
 
+        [JsonProperty("starsFive")]
+        public int StarsFive { get; set; }
+
+        [JsonProperty("starsFour")]
+        public int StarsFour { get; set; }
+
+        [JsonProperty("starsThree")]
+        public int StarsThree { get; set; }
+
+        [JsonProperty("starsTwo")]
+        public int StarsTwo { get; set; }
+
+        [JsonProperty("starsOne")]
+        public int StarsOne { get; set; }
+
         [JsonProperty("totalCount")]
         public long TotalCount { get; set; }
     }