+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- StarsFive, StarsFour, StarsThree, StarsTwo, StarsOne to the RatingResponse
+- 
 ## [3.17.0] - 2025-01-29